# server/member/api/openapi.yaml
openapi: 3.0.3
info:
  title: Member Service API
  version: 0.1.0
  description: 会員ドメイン。Step 0 の薄い骨格（liveness + 一覧）。
paths:
...
  /members/{id}/addresses:
    get:
      operationId: listAddresses
      summary: 会員の住所帳一覧
      tags: [address]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: 住所の配列
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createAddress
      summary: 住所を住所帳に追加
      tags: [address]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: 追加された住所
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        default:
          $ref: "#/components/responses/Error"
  /members/{id}/addresses/{addressId}:
    get:
      operationId: getAddress
      summary: 住所を取得
      tags: [address]
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/AddressIdPath"
      responses:
        "200":
          description: 住所
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        default:
          $ref: "#/components/responses/Error"
    put:
      operationId: updateAddress
      summary: 住所を更新
      tags: [address]
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/AddressIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "200":
          description: 更新された住所
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteAddress
      summary: 住所を削除
      tags: [address]
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/AddressIdPath"
      responses:
        "204":
          description: 削除済み (存在しない id でも冪等に 204)
        default:
          $ref: "#/components/responses/Error"
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: リソース ID
      schema:
        type: integer
        format: int64
    AddressIdPath:
      name: addressId
      in: path
      required: true
      description: 住所 ID
      schema:
        type: integer
        format: int64
...
  responses:
    Error:
      description: エラー (全サービス共通フォーマット — ErrorJSON ミドルウェアが整形)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: >-
            機械可読なエラー種別 (snake_case)。取りうる値は共通ミドルウェア
            server/internal/middleware (ErrorJSON / AppError) を唯一の出所とする。
          example: not_found
        message:
          type: string
          description: 人間可読なエラー説明 (内部詳細を含む 500 系は固定文言に伏せる)
          example: member 99 not found
...
    Address:
      type: object
      required: [id, memberId, recipient, postalCode, prefecture, city, line1, createdAt]
      properties:
        id:
          type: integer
          format: int64
        memberId:
          type: integer
          format: int64
        recipient:
          type: string
        postalCode:
          type: string
        prefecture:
          type: string
        city:
          type: string
        line1:
          type: string
        createdAt:
          type: string
          format: date-time
    AddressInput:
      type: object
      required: [recipient, postalCode, prefecture, city, line1]
      properties:
        recipient:
          type: string
          x-oapi-codegen-extra-tags:
            binding: required
        postalCode:
          type: string
          x-oapi-codegen-extra-tags:
            binding: required
        prefecture:
          type: string
          x-oapi-codegen-extra-tags:
            binding: required
        city:
          type: string
          x-oapi-codegen-extra-tags:
            binding: required
        line1:
          type: string
          x-oapi-codegen-extra-tags:
            binding: required